+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## unreleased
+### Added
+* Added an option to open the corresponding motion folder for a model folder from the Animations tab.
+
 ### Changed
 * Changed the Mesh Editor to use a side panel to select the entry to edit.
 * Changed the Hlpb Editor to use a side panel to select the entry to edit.